+++ src/plugins/rowSelector/index.ts
@@ -1,0 +1,4 @@
+export * from './basic/basicRowSelector.component';
+export * from './basic/basicRowSelector.interface';
+export * from './limited/limitedRowSelector.component';
+export * from './limited/limitedRowSelector.interface';