+++ changelog.md
@@ -52,6 +52,8 @@
 - renamed `BasicTableColumnContext` to `TableGridCellTemplateContext` and changed to interface
 - renamed `GridComponent` to `GridSAComponent`
 - renamed `SimpleNoDataRendererComponent` to `SimpleNoDataRendererSAComponent`
+- renamed `BasicRowSelectorComponent` to `BasicRowSelectorSAComponent`
+- renamed `LimitedRowSelectorComponent` to `LimitedRowSelectorSAComponent`
 - removed `GridPluginGeneric`, now `GridPlugin` is generic with default value
 - old `ContentRendererOptions` renamed to `HeaderBodyContentRendererOptions`
 - new `ContentRendererOptions` without `plugins`